+++ app/src/main/java/com/example/task_app/activities/CreateAccount.java
@@ -1,4 +1,4 @@
-package com.example.task_app;
+package com.example.task_app.activities;
 
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -7,7 +7,8 @@ import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
 
-import com.example.task_app.databinding.ActivityCreateAccountBinding;
+import com.example.task_app.R;
+
 
 public class CreateAccount extends AppCompatActivity {
