+++ app/src/main/java/com/example/task_app/activities/Pantalla_carga.java
@@ -1,4 +1,4 @@
-package com.example.task_app;
+package com.example.task_app.activities;
 
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -6,22 +6,27 @@ import android.content.Intent;
 import android.os.Bundle;
 import android.os.Handler;
 
-public class SyncUp extends AppCompatActivity {
+import com.example.task_app.MainActivity;
+import com.example.task_app.R;
+
+public class Pantalla_carga extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_sync_up);
+        setContentView(R.layout.activity_pantalla_carga);
 
         int Tiempo = 1000;
 
+        /**
+         * Pantalla de carga de la aplicacion
+         * */
         new Handler().postDelayed(new Runnable() {
             @Override
             public void run() {
-                startActivity(new Intent(SyncUp.this, LogIn.class));
+                startActivity(new Intent(Pantalla_carga.this, MainActivity.class));
                 finish();
             }
         },Tiempo);
-
     }
 }
